Extracted FirstJoinTask from AnalysisPipeline.Run
* New OspreySharp/Tasks/FirstJoinTask.cs owns Stage 5 (first-pass
  Percolator + first-pass protein FDR + 1st-pass FDR sidecar
  persistence + post-FDR compaction), the --join-at-pass=2 2nd-pass
  sidecar overlay, and the Stage 6 planning checkpoint
  (multi-charge consensus + cross-run RT consensus + per-file
  calibration refit + reconciliation planning + WriteReconciliationFiles)
* Outputs (PerFileConsensusTargets, ReconciliationActions,
  RefinedCalibrations, PerFileGapFillForRescore, DidPlan) flow to
  PerFileRescoreTask through instance properties on the task
* Changed OspreyTask.Run to return bool to support early-exit paths
  (StopAfterStage5 success, sidecar load failures); Pipeline.Execute
  short-circuits and returns false on the first task that does so.
  PipelineContext.ExitCode carries the requested process exit code.
  Mirrors the boolean-return pattern in
  pwiz_tools/Skyline/CommandLine.cs
* WriteReconciliationFiles, RunFdr, RunFirstPassProteinFdr, and
  WriteFdrScoresSidecars become internal so the new task can drive
  them; method bodies stay on AnalysisPipeline for now
* AnalysisPipeline.Run drops 582 lines of inline orchestration in
  favor of a 46-line task construction + dispatch
* Stellar 3-file snapshot regression PASS at every stage

See ai/todos/active/TODO-20260509_osprey_pipeline_tasks.md
