+++ pwiz_tools/OspreySharp/OspreySharp.Tasks/PipelineContext.cs
@@ -60,6 +60,18 @@ namespace pwiz.OspreySharp.Tasks
         /// </summary>
         public OspreyConfig Config { get; }
 
+        /// <summary>
+        /// Process exit code requested by a task that has returned
+        /// <c>false</c> from <see cref="OspreyTask.Run"/>. Defaults
+        /// to 0; tasks that short-circuit the pipeline because of an
+        /// error must set this to a non-zero value before returning.
+        /// The pipeline driver itself does not read this — it is
+        /// purely a channel for the caller of
+        /// <see cref="Pipeline.Execute"/> to learn what code the
+        /// pipeline wants the process to exit with.
+        /// </summary>
+        public int ExitCode { get; set; }
+
         public PipelineContext(OspreyConfig config,
             Action<string> logInfo,
             Action<string> logWarning,
